// src/api/schemas.gen.ts
export const BehaviorPolicySchema = {
    type: 'object',
    description: 'A policy which is used to validate and execute certain actions based on the validation result.',
    properties: {
        behavior: {
            $ref: '#/components/schemas/BehaviorPolicyBehavior'
        },
        createdAt: {
            type: 'string',
            format: 'date-time',
            description: 'The formatted UTC timestamp indicating when the policy was created.',
            readOnly: true
        },
        deserialization: {
            $ref: '#/components/schemas/BehaviorPolicyDeserialization'
        },
        id: {
            type: 'string',
            description: 'The unique identifier of the policy.'
        },
        lastUpdatedAt: {
            type: 'string',
            format: 'date-time',
            description: 'The formatted UTC timestamp indicating when the policy was updated the last time.',
            readOnly: true
        },
        matching: {
            $ref: '#/components/schemas/BehaviorPolicyMatching'
        },
        onTransitions: {
            type: 'array',
            items: {
                $ref: '#/components/schemas/BehaviorPolicyOnTransition'
            }
        }
    },
    required: [
        'behavior',
        'id',
        'matching'
    ]
} as const;
// ...
export const DataPolicySchema = {
    type: 'object',
    description: 'A data policy which is used to validate and execute certain actions based on the validation result.',
    properties: {
        createdAt: {
            type: 'string',
            format: 'date-time',
            description: 'The formatted UTC timestamp indicating when the policy was created.',
            readOnly: true
        },
        id: {
            type: 'string',
            description: 'The unique identifier of the policy.'
        },
        lastUpdatedAt: {
            type: 'string',
            format: 'date-time',
            description: 'The formatted UTC timestamp indicating when the policy was updated the last time.',
            readOnly: true
        },
        matching: {
            $ref: '#/components/schemas/DataPolicyMatching'
        },
        onFailure: {
            $ref: '#/components/schemas/DataPolicyAction'
        },
        onSuccess: {
            $ref: '#/components/schemas/DataPolicyAction'
        },
        validation: {
            $ref: '#/components/schemas/DataPolicyValidation'
        }
    },
    required: [
        'id',
        'matching'
    ]
} as const;
// ...
export const PolicySchemaSchema = {
    type: 'object',
    properties: {
        arguments: {
            type: 'object',
            additionalProperties: {
                type: 'string',
                description: 'The schema type dependent arguments.'
            },
            description: 'The schema type dependent arguments.'
        },
        createdAt: {
            type: 'string',
            description: 'The formatted UTC timestamp when the schema was created.',
            readOnly: true
        },
        id: {
            type: 'string',
            description: 'The unique identifier of the schema.'
        },
        schemaDefinition: {
            type: 'string',
            description: 'The base64 encoded schema definition.'
        },
        type: {
            type: 'string',
            description: 'The type of the schema.'
        },
        version: {
            type: 'integer',
            format: 'int32',
            description: 'The version of the schema.',
            readOnly: true
        }
    },
    required: [
        'id',
        'schemaDefinition',
        'type'
    ]
} as const;
// ...
export const ScriptSchema = {
    type: 'object',
    properties: {
        createdAt: {
            type: 'string',
            description: 'The formatted UTC timestamp when the script was created.',
            readOnly: true
        },
        description: {
            type: 'string',
            description: 'A string of free-form text describing the function.'
        },
        functionType: {
            type: 'string',
            description: 'The type of the function.',
            enum: [
                'TRANSFORMATION'
            ]
        },
        id: {
            type: 'string',
            description: 'The unique identifier of the script.'
        },
        source: {
            type: 'string',
            description: 'The base64 encoded function source code.'
        },
        version: {
            type: 'integer',
            format: 'int32',
            description: 'The version of the script.',
            readOnly: true
        }
    },
    required: [
        'functionType',
        'id',
        'source'
    ]
} as const;
// ...
export const BridgeSchema = {
    type: 'object',
    properties: {
        cleanStart: {
            type: 'boolean',
            format: 'boolean',
            default: true,
            description: 'The cleanStart value associated the the MQTT connection.'
        },
        clientId: {
            type: 'string',
            format: 'string',
            description: 'The client identifier associated the the MQTT connection.',
            example: 'my-example-client-id',
            maxLength: 65535,
            nullable: true
        },
        host: {
            type: 'string',
            description: 'The host the bridge connects to - a well formed hostname, ipv4 or ipv6 value.',
            maxLength: 255
        },
        id: {
            type: 'string',
            format: 'string',
            description: 'The bridge id, must be unique and only contain alpha numeric characters with spaces and hyphens.',
            maxLength: 500,
            minLength: 1,
            pattern: '^([a-zA-Z_0-9-_])*$'
        },
        keepAlive: {
            type: 'integer',
            format: 'int32',
            default: 240,
            description: 'The keepAlive associated the the MQTT connection.',
            maximum: 65535,
            minimum: 0
        },
        localSubscriptions: {
            type: 'array',
            description: 'localSubscriptions associated with the bridge',
            items: {
                $ref: '#/components/schemas/LocalBridgeSubscription'
            }
        },
        loopPreventionEnabled: {
            type: 'boolean',
            format: 'boolean',
            default: true,
            description: 'Is loop prevention enabled on the connection'
        },
        loopPreventionHopCount: {
            type: 'integer',
            format: 'int32',
            default: 1,
            description: 'Loop prevention hop count',
            maximum: 100,
            minimum: 0
        },
        password: {
            type: 'string',
            format: 'string',
            description: 'The password value associated the the MQTT connection.',
            maxLength: 65535,
            nullable: true
        },
        persist: {
            type: 'boolean',
            description: 'If this flag is set to true, any outgoing mqtt messages with QoS-1 or QoS-2 will be persisted on disc in case disc persistence is active.If this flag is set to false, the QoS of any outgoing mqtt messages will be set to QoS-0 and no traffic will be persisted on disc.',
            nullable: true
        },
        port: {
            type: 'integer',
            format: 'int32',
            description: 'The port number to connect to',
            maximum: 65535,
            minimum: 1
        },
        remoteSubscriptions: {
            type: 'array',
            description: 'remoteSubscriptions associated with the bridge',
            items: {
                $ref: '#/components/schemas/BridgeSubscription'
            }
        },
        sessionExpiry: {
            type: 'integer',
            format: 'int64',
            default: 3600,
            description: 'The sessionExpiry associated the the MQTT connection.',
            minimum: 0
        },
        status: {
            $ref: '#/components/schemas/Status'
        },
        tlsConfiguration: {
            $ref: '#/components/schemas/TlsConfiguration'
        },
        username: {
            type: 'string',
            format: 'string',
            description: 'The username value associated the the MQTT connection.',
            maxLength: 65535,
            nullable: true
        },
        websocketConfiguration: {
            $ref: '#/components/schemas/WebsocketConfiguration'
        }
    },
    required: [
        'cleanStart',
        'host',
        'id',
        'keepAlive',
        'port',
        'sessionExpiry'
    ]
} as const;
// ...
export const StatusTransitionCommandSchema = {
    type: 'object',
    properties: {
        command: {
            type: 'string',
            description: 'The command to perform on the target connection.',
            enum: [
                'START',
                'STOP',
                'RESTART'
            ]
        }
    }
} as const;
// ...
export const AdapterSchema = {
    type: 'object',
    properties: {
        config: {
            $ref: '#/components/schemas/JsonNode'
        },
        id: {
            type: 'string',
            format: 'string',
            description: 'The adapter id, must be unique and only contain alpha numeric characters with spaces and hyphens.',
            maxLength: 500,
            minLength: 1,
            pattern: '^([a-zA-Z_0-9-_])*$'
        },
        status: {
            $ref: '#/components/schemas/Status'
        },
        type: {
            type: 'string',
            description: 'The adapter type associated with this instance'
        }
    },
    required: [
        'id'
    ]
} as const;